Fix: the 2D mapper no longer freezes on a map holding a room at the coordinate limit
QA finding F-17-1 (alias F-R9-1). Regression from PR #10175 (714efcaf9),
which introduced TAreaGridIndex and the roomsInViewport scan.

The viewport queries have two ways through a coordinate range: probe the
range's keys one at a time, or walk the keys the level holds and drop the
ones outside it. The probing loops counted in int, so a range ending at
INT_MAX never got past its last step - `for (int x = minX; x <= maxX; ++x)`
has no value greater than maxX to stop at, and the counter wrapped back to
INT_MIN and climbed again, forever. The paint never returned, so the client
froze at 100% CPU with no way back but replacing the map file.

Two ordinary routes reach such a range. One room at x = 2147483647 in an
area a few hundred columns wide, centred on and drawn at one pixel per room
- zoom 400, which the scroll wheel reaches - puts the viewport's right-hand
bound exactly on the limit, and the freeze then travels inside the saved
map into the next session's loadMap(). Separately, a zoom far enough out to
overflow the float span leaves a room zero pixels wide, so the inverse
transform divides by that zero. What comes back depends on the platform - a
NaN where the panned offset converts to 0, an infinity where it converts to
INT_MIN - and neither survives qBound as a range: a NaN settles both bounds on
the lower limit, an infinity pins both on the upper one, so the viewport
collapses onto one end of the coordinate space either way.

The loops now count in qint64, which terminates for every int range
including [INT_MIN, INT_MAX] on both axes, and viewportRoomBounds() names
the end of the range that stands in for a NaN, so an absurd zoom widens the
viewport to everything and draws every room into the one pixel the map has
become - what 5.0.1 did - instead of collapsing it onto the limit.

TAreaGridIndexTest gains seven cases at the coordinate limits, and is
capped at 60s in ctest because a regression here hangs rather than fails.
MapViewportLimits_spec.lua covers both user routes through the real widget.

+++ src/TAreaGridIndex.cpp
@@ -26,12 +26,16 @@ const TAreaGridIndex::RoomIds TAreaGridIndex::csmEmptyCell;
 // it. A zoomed-in viewport covers a handful of the rows a level occupies and a
 // zoomed-out one covers more rows than exist, so neither wins outright - the
 // smaller key count is the one to get through.
+//
+// The probing loops count in qint64 rather than int: a viewport can reach the
+// coordinate limit, and an int counter at INT_MAX never compares greater than
+// it - it overflows back to INT_MIN and the loop runs forever.
 static void appendRoomsInYRange(const QHash<int, TAreaGridIndex::RoomIds>& yMap, int minY, int maxY, QList<int>& result)
 {
     const qint64 wantedRows = static_cast<qint64>(maxY) - minY + 1;
     if (wantedRows > 0 && wantedRows < yMap.size()) {
-        for (int y = minY; y <= maxY; ++y) {
-            const auto yIt = yMap.constFind(y);
+        for (qint64 y = minY; y <= maxY; ++y) {
+            const auto yIt = yMap.constFind(static_cast<int>(y));
             if (yIt == yMap.constEnd()) {
                 continue;
             }
@@ -55,8 +59,8 @@ static void appendRoomsInYRangeWithCollisions(const QHash<int, TAreaGridIndex::R
 {
     const qint64 wantedRows = static_cast<qint64>(maxY) - minY + 1;
     if (wantedRows > 0 && wantedRows < yMap.size()) {
-        for (int y = minY; y <= maxY; ++y) {
-            const auto yIt = yMap.constFind(y);
+        for (qint64 y = minY; y <= maxY; ++y) {
+            const auto yIt = yMap.constFind(static_cast<int>(y));
             if (yIt == yMap.constEnd()) {
                 continue;
             }
@@ -190,8 +194,8 @@ QList<int> TAreaGridIndex::roomsInViewport(int z, int minX, int maxX, int minY, 
     const auto& xyMap = *zIt;
     const qint64 wantedColumns = static_cast<qint64>(maxX) - minX + 1;
     if (wantedColumns > 0 && wantedColumns < xyMap.size()) {
-        for (int x = minX; x <= maxX; ++x) {
-            const auto xIt = xyMap.constFind(x);
+        for (qint64 x = minX; x <= maxX; ++x) {
+            const auto xIt = xyMap.constFind(static_cast<int>(x));
             if (xIt == xyMap.constEnd()) {
                 continue;
             }
@@ -218,8 +222,8 @@ QList<QPair<int, bool>> TAreaGridIndex::roomsInViewportWithCollisions(int z, int
     const auto& xyMap = *zIt;
     const qint64 wantedColumns = static_cast<qint64>(maxX) - minX + 1;
     if (wantedColumns > 0 && wantedColumns < xyMap.size()) {
-        for (int x = minX; x <= maxX; ++x) {
-            const auto xIt = xyMap.constFind(x);
+        for (qint64 x = minX; x <= maxX; ++x) {
+            const auto xIt = xyMap.constFind(static_cast<int>(x));
             if (xIt == xyMap.constEnd()) {
                 continue;
             }
